requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib[argon2]
python-jose[cryptography]
python-multipart
pytest
pytest-asyncio
httpx
mongomock
pytest-mock
pytest-cov